+++ Source/TRONCodable.swift
@@ -71,7 +71,7 @@ open class CodableSerializer {
 
      - returns: APIRequest instance.
      */
-    public func request<Model: Decodable, ErrorModel: ErrorSerializable>(_ path: String) -> APIRequest<Model, ErrorModel> {
+    open func request<Model: Decodable, ErrorModel: ErrorSerializable>(_ path: String) -> APIRequest<Model, ErrorModel> {
         return tron.request(path, responseSerializer: CodableParser(modelDecoder: modelDecoder))
     }
 
@@ -84,7 +84,7 @@ open class CodableSerializer {
 
      - returns: APIRequest instance.
      */
-    public func upload<Model: Decodable, ErrorModel: ErrorSerializable>(_ path: String, fromFileAt fileURL: URL) -> UploadAPIRequest<Model, ErrorModel> {
+    open func upload<Model: Decodable, ErrorModel: ErrorSerializable>(_ path: String, fromFileAt fileURL: URL) -> UploadAPIRequest<Model, ErrorModel> {
         return tron.upload(path, fromFileAt: fileURL,
                            responseSerializer: CodableParser(modelDecoder: modelDecoder))
     }
@@ -98,7 +98,7 @@ open class CodableSerializer {
 
      - returns: APIRequest instance.
      */
-    public func upload<Model: Decodable, ErrorModel: ErrorSerializable>(_ path: String, data: Data) -> UploadAPIRequest<Model, ErrorModel> {
+    open func upload<Model: Decodable, ErrorModel: ErrorSerializable>(_ path: String, data: Data) -> UploadAPIRequest<Model, ErrorModel> {
         return tron.upload(path, data: data, responseSerializer: CodableParser(modelDecoder: modelDecoder))
     }
 
@@ -111,7 +111,7 @@ open class CodableSerializer {
 
      - returns: APIRequest instance.
      */
-    public func upload<Model: Decodable, ErrorModel: ErrorSerializable>(_ path: String, from stream: InputStream) -> UploadAPIRequest<Model, ErrorModel> {
+    open func upload<Model: Decodable, ErrorModel: ErrorSerializable>(_ path: String, from stream: InputStream) -> UploadAPIRequest<Model, ErrorModel> {
         return tron.upload(path, from: stream, responseSerializer: CodableParser(modelDecoder: modelDecoder))
     }
 
@@ -124,7 +124,7 @@ open class CodableSerializer {
 
      - returns: MultipartAPIRequest instance.
      */
-    public func uploadMultipart<Model: Decodable, ErrorModel: ErrorSerializable>(_ path: String,
+    open func uploadMultipart<Model: Decodable, ErrorModel: ErrorSerializable>(_ path: String,
                                                                                  encodingMemoryThreshold: UInt64 = MultipartUpload.encodingMemoryThreshold,
                                                                                  fileManager: FileManager = .default,
                                                                                  formData: @escaping (MultipartFormData) -> Void) -> UploadAPIRequest<Model, ErrorModel> {